Update main_export with no_dynamic_axes=True and update examples
